Fix links to tags and categories
The stand-alone category and tag pages were not being generated
so the tab and category links in individual distro pages
(which pointed to pages like /tags/server or /category/beginner)
resulted in 404 errors.

In the case of categories, the individual category pages were merely lists of distros in them,
in exactly the same format as what's shown in at /categories (but divided by category).
A simpler approach is then to simply link each category
to the corresponding section of the /categories page.

As for the tags, there used to be a different type of overview at /tags,
but that was just a list of tags with counts of distros within them
(see archived version here: https://web.archive.org/web/20210921104837/https://computefreely.org/tags/).
An actual listing of distros per tag is more useful,
and more consistent with the categories overview (and simpler to implement),
so the category overview page was duplicated and adapted to present a tags overview,
which then also allows having each tag linking to the corresponding section in that page.

With this simplification we are also able to remove the unused layouts for the individual tag and category pages,
as well as the plugins to generate those pages (which aren't working anyway, at least not in GitHub Pages).